backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
python-dotenv==1.0.0
Werkzeug==2.3.7
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
Authlib==1.2.1
requests==2.31.0
Pillow==10.0.0
gunicorn==21.2.0
python-dateutil==2.8.2
cryptography==41.0.4
bcrypt==4.0.1
marshmallow==3.20.1
pandas==2.1.0